Split metadata mutation tracking
