+++ src/app/controller/library/wavs/metadata_async.rs
@@ -7,11 +7,12 @@ use crate::app::controller::jobs::{
 };
 use crate::app::controller::state::runtime::{MetadataRollback, PendingMetadataMutation};
 use std::collections::BTreeSet;
-use std::path::{Path, PathBuf};
+use std::path::PathBuf;
 use std::time::{Duration, Instant};
 
 mod analysis_ops;
 mod source_ops;
+mod tracking;
 
 const SELECTED_SOURCE_MUTATION_CLAIM_GRACE: Duration = Duration::from_millis(750);
 const SELECTED_SOURCE_MUTATION_AUTO_SYNC_GRACE: Duration = Duration::from_secs(5);
@@ -19,57 +20,6 @@ const METADATA_FILE_OP_PRIORITY_WAIT_TIMEOUT: Duration = Duration::from_secs(8);
 const METADATA_FILE_OP_PRIORITY_WAIT_DELAY: Duration = Duration::from_millis(100);
 
 impl AppController {
-    pub(crate) fn selected_source_claim_pause_grace_active(&mut self, now: Instant) -> bool {
-        let Some(source_id) = self.selected_source_id() else {
-            return false;
-        };
-        self.runtime
-            .source_lane
-            .mutations
-            .claim_pause_grace_active(&source_id, now)
-    }
-
-    pub(crate) fn extend_selected_source_mutation_claim_grace(&mut self, source_id: &SourceId) {
-        if self.selected_source_id().as_ref() != Some(source_id) {
-            return;
-        }
-        self.runtime.source_lane.mutations.extend_claim_pause_grace(
-            source_id,
-            Instant::now() + SELECTED_SOURCE_MUTATION_CLAIM_GRACE,
-        );
-    }
-
-    pub(crate) fn selected_source_auto_sync_grace_active(&mut self, now: Instant) -> bool {
-        let Some(source_id) = self.selected_source_id() else {
-            return false;
-        };
-        self.runtime
-            .source_lane
-            .mutations
-            .auto_sync_grace_active(&source_id, now)
-    }
-
-    pub(crate) fn source_auto_sync_grace_active(
-        &mut self,
-        source_id: &SourceId,
-        now: Instant,
-    ) -> bool {
-        self.runtime
-            .source_lane
-            .mutations
-            .auto_sync_grace_active(source_id, now)
-    }
-
-    pub(crate) fn extend_selected_source_mutation_auto_sync_grace(&mut self, source_id: &SourceId) {
-        if self.selected_source_id().as_ref() != Some(source_id) {
-            return;
-        }
-        self.runtime.source_lane.mutations.extend_auto_sync_grace(
-            source_id,
-            Instant::now() + SELECTED_SOURCE_MUTATION_AUTO_SYNC_GRACE,
-        );
-    }
-
     /// Queue one metadata mutation batch on a worker thread after optimistic UI updates.
     pub(crate) fn queue_metadata_mutation(
         &mut self,
@@ -170,38 +120,6 @@ impl AppController {
         );
     }
 
-    /// Return whether one sample path already has an optimistic metadata write in flight.
-    pub(crate) fn metadata_mutation_pending_for(
-        &self,
-        source_id: &SourceId,
-        relative_path: &Path,
-    ) -> bool {
-        self.runtime
-            .source_lane
-            .mutations
-            .metadata_path_pending(source_id, relative_path)
-    }
-
-    /// Return whether the currently selected source still has optimistic metadata writes pending.
-    pub(crate) fn selected_source_has_pending_metadata_mutations(&self) -> bool {
-        self.selected_source_id().is_some_and(|source_id| {
-            self.runtime
-                .source_lane
-                .mutations
-                .source_has_pending_metadata(&source_id)
-        })
-    }
-
-    /// Return whether the currently selected source still owns a background file mutation.
-    pub(crate) fn selected_source_has_pending_file_mutations(&self) -> bool {
-        self.selected_source_id().is_some_and(|source_id| {
-            self.runtime
-                .source_lane
-                .mutations
-                .source_has_pending_file_mutations(&source_id)
-        })
-    }
-
     /// Return whether any streamed file operation is currently in progress.
     pub(crate) fn file_ops_in_progress_for_projection(&self) -> bool {
         self.runtime.jobs.file_ops_in_progress()
@@ -210,62 +128,6 @@ impl AppController {
     /// Return whether a waveform image render is still in flight.
     pub(crate) fn waveform_render_in_progress_for_projection(&self) -> bool {
         self.runtime.pending_waveform_render.is_some()
-    }
-
-    /// Return whether one source currently owns a background file mutation.
-    pub(crate) fn source_has_pending_file_mutations(&self, source_id: &SourceId) -> bool {
-        self.runtime
-            .source_lane
-            .mutations
-            .source_has_pending_file_mutations(source_id)
-    }
-
-    /// Mark one source/path batch as owned by a background file mutation.
-    pub(crate) fn begin_pending_file_mutation(
-        &mut self,
-        source_id: &SourceId,
-        paths: impl IntoIterator<Item = PathBuf>,
-    ) {
-        let paths = paths.into_iter().collect::<Vec<_>>();
-        let source_became_active = self
-            .runtime
-            .source_lane
-            .mutations
-            .begin_file_mutation(source_id, paths.clone());
-        if source_became_active {
-            crate::app::controller::library::source_write_priority::begin_file_op_write_priority(
-                source_id,
-            );
-        }
-        self.runtime
-            .jobs
-            .begin_source_watch_file_op(source_id.clone(), paths);
-        self.extend_selected_source_mutation_claim_grace(source_id);
-        self.extend_selected_source_mutation_auto_sync_grace(source_id);
-    }
-
-    /// Clear one source/path batch from background file-mutation tracking.
-    pub(crate) fn finish_pending_file_mutation(
-        &mut self,
-        source_id: &SourceId,
-        paths: impl IntoIterator<Item = PathBuf>,
-    ) {
-        let paths = paths.into_iter().collect::<Vec<_>>();
-        let source_became_inactive = self
-            .runtime
-            .source_lane
-            .mutations
-            .finish_file_mutation(source_id, paths.clone());
-        if source_became_inactive {
-            crate::app::controller::library::source_write_priority::finish_file_op_write_priority(
-                source_id,
-            );
-        }
-        self.runtime
-            .jobs
-            .finish_source_watch_file_op(source_id.clone(), paths);
-        self.extend_selected_source_mutation_claim_grace(source_id);
-        self.extend_selected_source_mutation_auto_sync_grace(source_id);
     }
 }
 
@@ -343,6 +205,7 @@ mod tests {
     use super::*;
     use crate::app::controller::library::source_write_priority;
     use std::io;
+    use std::path::Path;
     use std::sync::{Arc, LazyLock, Mutex, MutexGuard};
     use tracing_subscriber::fmt::MakeWriter;
